Prevent project viewer from creating cluster templates (#7446) (#7482)
* Fix permission config key for cluster template obj

* Refactor code to remove redundancy and display tooltip in disabled state

* Improve Readability of code to allow deletion of template if user is is owner itself

* Handle Avoid deletion of clustertemplates when role as a viewer

* Update Error/Test Messages for viewer role

(cherry picked from commit 9c39de9a1bdab072a12a559f81d13cbec417d1a3)

+++ modules/api/pkg/provider/kubernetes/cluster_template.go
@@ -67,8 +67,8 @@ func (p *ClusterTemplateProvider) CreateorUpdate(ctx context.Context, userInfo *
 		return nil, errors.New("the global scope is reserved only for admins")
 	}
 
-	if userInfo.Roles.Has("viewers") && userInfo.Roles.Len() == 1 && scope != kubermaticv1.UserClusterTemplateScope {
-		return nil, fmt.Errorf("viewer is not allowed to create cluster template for the %s scope", scope)
+	if userInfo.Roles.Has("viewers") && userInfo.Roles.Len() == 1 {
+		return nil, utilerrors.New(http.StatusForbidden, fmt.Sprintf("user %s has viewer role and cannot create or update cluster templates regardless of any scope", userInfo.Email))
 	}
 
 	if scope == kubermaticv1.ProjectClusterTemplateScope && projectID == "" {
@@ -178,6 +178,10 @@ func (p *ClusterTemplateProvider) Delete(ctx context.Context, userInfo *provider
 	// only admin can delete global templates
 	if !userInfo.IsAdmin && result.Labels[kubermaticv1.ClusterTemplateScopeLabelKey] == kubermaticv1.GlobalClusterTemplateScope {
 		return utilerrors.New(http.StatusForbidden, fmt.Sprintf("user %s can't delete template %s", userInfo.Email, templateID))
+	}
+
+	if userInfo.Roles.Has("viewers") && userInfo.Roles.Len() == 1 {
+		return utilerrors.New(http.StatusForbidden, fmt.Sprintf("user %s has viewer role and cannot delete cluster templates regardless of any scope", userInfo.Email))
 	}
 
 	return p.clientPrivileged.Delete(ctx, result)
